11 HW: cache. DB VS DB+CACHE. corrections
